# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/v1/models": {
      "get": {
        "summary": "获取模型列表",
        "deprecated": false,
        "description": "获取当前可用的模型列表。\n\n根据请求头自动识别返回格式：\n- 包含 `x-api-key` 和 `anthropic-version` 头时返回 Anthropic 格式\n- 包含 `x-goog-api-key` 头或 `key` 查询参数时返回 Gemini 格式\n- 其他情况返回 OpenAI 格式\n",
        "operationId": "listModels",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "Anthropic API Key (用于 Claude 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "anthropic-version",
            "in": "header",
            "description": "Anthropic API 版本",
            "required": false,
            "example": "",
            "schema": {
              "type": "string",
              "example": "2023-06-01"
            }
          },
          {
            "name": "x-goog-api-key",
            "in": "header",
            "description": "Google API Key (用于 Gemini 格式)",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1beta/models": {
      "get": {
        "summary": "Gemini 格式获取",
        "deprecated": false,
        "description": "以 Gemini API 格式返回可用模型列表",
        "operationId": "listModelsGemini",
        "tags": [
          "获取模型列表"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeminiModelsResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
	    "/v1/responses/compact": {
	      "post": {
	        "summary": "压缩对话 (OpenAI Responses API)",
	        "deprecated": false,
	        "description": "OpenAI Responses API，用于对长对话进行 compaction。",
	        "operationId": "compactResponse",
        "tags": [
          "OpenAI格式(Responses)"
        ],
        "parameters": [],
	        "requestBody": {
	          "content": {
	            "application/json": {
	              "schema": {
	                "$ref": "#/components/schemas/ResponsesCompactionRequest"
	              }
	            }
	          },
	          "required": true
	        },
        "responses": {
          "200": {
            "description": "成功压缩对话",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesCompactionResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/images/generations": {
      "post": {
        "summary": "生成图像",
        "deprecated": false,
        "description": "图片生成。默认同步返回图片结果；请求体设置 `async: true` 或发送 `Prefer: respond-async` 可请求异步处理。异步模式支持阿里原生异步图片模型，以及 Grsai 的 nano-banana / gpt-image-2 图片模型。Grsai 渠道通过官方 Base URL 自动识别，也可在渠道设置中指定 `async_image_provider: grsai`；Grsai 异步请求当前仅支持 `n=1`。所有异步请求均要求 `response_format` 为 `url`、`stream` 不为 `true`。计费模式为 `tiered_expr` 的模型当前不支持异步图片请求，此类请求返回 HTTP 400。",
        "operationId": "createImage",
        "tags": [
          "图片生成/Qwen千问"
        ],
        "parameters": [
          {
            "name": "Prefer",
            "in": "header",
            "description": "包含 `respond-async` 指令时请求异步处理，与请求体 `async: true` 等效。",
            "required": false,
            "example": "respond-async",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "async": {
                    "type": "boolean",
                    "description": "是否异步处理。设为 `true` 时成功提交返回 HTTP 202 和图片任务对象；默认仍按同步方式处理。支持阿里原生异步图片模型和 Grsai nano-banana / gpt-image-2 图片模型；Grsai 当前仅支持 `n=1`。`tiered_expr` 计费模型当前不支持异步处理并返回 HTTP 400。",
                    "default": false
                  },
                  "input": {
                    "type": "object",
                    "properties": {
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "role": {
                              "type": "string"
                            },
                            "content": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "text": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "messages"
                    ]
                  },
                  "parameters": {
                    "type": "object",
                    "properties": {
                      "negative_prompt": {
                        "type": "string"
                      },
                      "prompt_extend": {
                        "type": "boolean"
                      },
                      "watermark": {
                        "type": "boolean"
                      },
                      "size": {
                        "type": "string"
                      }
                    }
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              },
              "example": {
                "model": "qwen-image-plus",
                "input": {
                  "messages": [
                    {
                      "role": "user",
                      "content": [
                        {
                          "text": "一副典雅庄重的对联悬挂于厅堂之中，房间是个安静古典的中式布置，桌子上放着一些青花瓷，对联上左书“义本生知人机同道善思新”，右书“通云赋智乾坤启数高志远”， 横批“智启通义”，字体飘逸，在中间挂着一幅中国风的画作，内容是岳阳楼。"
                        }
                      ]
                    }
                  ]
                },
                "parameters": {
                  "negative_prompt": "",
                  "prompt_extend": true,
                  "watermark": false,
                  "size": "1328*1328"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功生成图像",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            },
            "headers": {}
          },
          "202": {
            "description": "异步图片生成任务已接受",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageTaskResponse"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "图片任务状态查询路径",
                "schema": {
                  "type": "string",
                  "format": "uri-reference",
                  "example": "/v1/images/generations/task_abc123"
                }
              },
              "Retry-After": {
                "description": "建议再次查询前等待的秒数",
                "schema": {
                  "type": "string",
                  "example": "5"
                }
              },
              "Preference-Applied": {
                "description": "使用 `Prefer: respond-async` 请求异步处理时返回",
                "schema": {
                  "type": "string",
                  "example": "respond-async"
                }
              }
            }
          },
          "400": {
            "description": "异步请求参数或计费模式不受支持。`tiered_expr` 计费模型发起异步图片请求时返回此状态。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                },
                "example": {
                  "code": "async_image_billing_not_supported",
                  "message": "tiered expression billing is not supported for asynchronous image generation",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/video/generations": {
      "post": {
        "summary": "创建视频生成任务",
        "deprecated": false,
        "description": "提交视频生成任务，支持文生视频和图生视频。\n\n返回任务 ID，可通过 GET 接口查询任务状态。\n",
        "operationId": "createVideoGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "kling-v1",
                "prompt": "宇航员在月球上漫步",
                "duration": 5,
                "width": 1280,
                "height": 720
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建视频生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/audio/speech": {
      "post": {
        "summary": "文本转语音",
        "deprecated": false,
        "description": "将文本转换为音频",
        "operationId": "createSpeech",
        "tags": [
          "OpenAI音频(Audio)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功生成音频",
            "content": {
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModelsResponse": {
        "type": "object",
        "properties": {
          "object": {
            "type": "string",
            "example": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
...
      "VideoResponse": {
        "type": "object",
        "description": "视频生成任务提交响应",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "任务 ID",
            "example": "abcd1234efgh"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "example": "queued"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}